+++ scalr/data_source_endpoint_test.go
@@ -2,15 +2,13 @@ package scalr
 
 import (
 	"fmt"
-	"math/rand"
 	"testing"
-	"time"
 
 	"github.com/hashicorp/terraform/helper/resource"
 )
 
 func TestAccEndpointDataSource_basic(t *testing.T) {
-	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	rInt := GetRandomInteger()
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:  func() { testAccPreCheck(t) },
@@ -43,7 +41,7 @@ resource scalr_environment test {
   name       = "test-env-%[1]d"
   account_id = "%s"
 }
-  
+
 resource scalr_endpoint test {
   name         = "test endpoint-%[1]d"
   timeout      = 15