+++ scalr/resource_scalr_workspace_test.go
@@ -4,9 +4,7 @@ import (
 	"context"
 	"fmt"
 	"log"
-	"math/rand"
 	"testing"
-	"time"
 
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
@@ -15,7 +13,7 @@ import (
 
 func TestAccScalrWorkspace_basic(t *testing.T) {
 	workspace := &scalr.Workspace{}
-	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	rInt := GetRandomInteger()
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
@@ -36,6 +34,14 @@ func TestAccScalrWorkspace_basic(t *testing.T) {
 						"scalr_workspace.test", "operations", "true"),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "working_directory", ""),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.pre_plan", "./scripts/pre-plan.sh"),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.post_plan", "./scripts/post-plan.sh"),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.pre_apply", "./scripts/pre-apply.sh"),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.post_apply", "./scripts/post-apply.sh"),
 					resource.TestCheckResourceAttrSet("scalr_workspace.test", "created_by.0.full_name"),
 					resource.TestCheckResourceAttrSet("scalr_workspace.test", "created_by.0.email"),
 					resource.TestCheckResourceAttrSet("scalr_workspace.test", "created_by.0.username"),
@@ -47,7 +53,7 @@ func TestAccScalrWorkspace_basic(t *testing.T) {
 
 func TestAccScalrWorkspace_monorepo(t *testing.T) {
 	workspace := &scalr.Workspace{}
-	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	rInt := GetRandomInteger()
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
@@ -74,7 +80,7 @@ func TestAccScalrWorkspace_monorepo(t *testing.T) {
 
 func TestAccScalrWorkspace_renamed(t *testing.T) {
 	workspace := &scalr.Workspace{}
-	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	rInt := GetRandomInteger()
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
@@ -95,6 +101,14 @@ func TestAccScalrWorkspace_renamed(t *testing.T) {
 						"scalr_workspace.test", "operations", "true"),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "working_directory", ""),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.pre_plan", "./scripts/pre-plan.sh"),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.post_plan", "./scripts/post-plan.sh"),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.pre_apply", "./scripts/pre-apply.sh"),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.post_apply", "./scripts/post-apply.sh"),
 				),
 			},
 
@@ -114,6 +128,14 @@ func TestAccScalrWorkspace_renamed(t *testing.T) {
 						"scalr_workspace.test", "operations", "true"),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "working_directory", ""),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.pre_plan", "./scripts/pre-plan.sh"),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.post_plan", "./scripts/post-plan.sh"),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.pre_apply", "./scripts/pre-apply.sh"),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.post_apply", "./scripts/post-apply.sh"),
 				),
 			},
 		},
@@ -121,7 +143,7 @@ func TestAccScalrWorkspace_renamed(t *testing.T) {
 }
 func TestAccScalrWorkspace_update(t *testing.T) {
 	workspace := &scalr.Workspace{}
-	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	rInt := GetRandomInteger()
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
@@ -131,17 +153,20 @@ func TestAccScalrWorkspace_update(t *testing.T) {
 			{
 				Config: testAccScalrWorkspaceBasic(rInt),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckScalrWorkspaceExists(
-						"scalr_workspace.test", workspace),
+					testAccCheckScalrWorkspaceExists("scalr_workspace.test", workspace),
 					testAccCheckScalrWorkspaceAttributes(workspace),
+					resource.TestCheckResourceAttr("scalr_workspace.test", "name", "workspace-test"),
+					resource.TestCheckResourceAttr("scalr_workspace.test", "auto_apply", "true"),
+					resource.TestCheckResourceAttr("scalr_workspace.test", "operations", "true"),
+					resource.TestCheckResourceAttr("scalr_workspace.test", "working_directory", ""),
 					resource.TestCheckResourceAttr(
-						"scalr_workspace.test", "name", "workspace-test"),
+						"scalr_workspace.test", "hooks.0.pre_plan", "./scripts/pre-plan.sh"),
 					resource.TestCheckResourceAttr(
-						"scalr_workspace.test", "auto_apply", "true"),
+						"scalr_workspace.test", "hooks.0.post_plan", "./scripts/post-plan.sh"),
 					resource.TestCheckResourceAttr(
-						"scalr_workspace.test", "operations", "true"),
+						"scalr_workspace.test", "hooks.0.pre_apply", "./scripts/pre-apply.sh"),
 					resource.TestCheckResourceAttr(
-						"scalr_workspace.test", "working_directory", ""),
+						"scalr_workspace.test", "hooks.0.post_apply", "./scripts/post-apply.sh"),
 				),
 			},
 
@@ -161,6 +186,24 @@ func TestAccScalrWorkspace_update(t *testing.T) {
 						"scalr_workspace.test", "terraform_version", "0.12.19"),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.test", "working_directory", "terraform/test"),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.pre_plan", "./scripts/pre-plan_updated.sh"),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.post_plan", "./scripts/post-plan_updated.sh"),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.pre_apply", "./scripts/pre-apply_updated.sh"),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "hooks.0.post_apply", "./scripts/post-apply_updated.sh"),
+				),
+			},
+
+			{
+				Config: testAccScalrWorkspaceUpdateWorkingDir(rInt),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckScalrWorkspaceExists(
+						"scalr_workspace.test", workspace),
+					resource.TestCheckResourceAttr(
+						"scalr_workspace.test", "working_directory", ""),
 				),
 			},
 		},
@@ -168,7 +211,7 @@ func TestAccScalrWorkspace_update(t *testing.T) {
 }
 
 func TestAccScalrWorkspace_import(t *testing.T) {
-	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	rInt := GetRandomInteger()
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
@@ -356,6 +399,12 @@ resource scalr_workspace test {
   name           = "workspace-test"
   environment_id = scalr_environment.test.id
   auto_apply     = true
+  hooks {
+    pre_plan   = "./scripts/pre-plan.sh"
+    post_plan  = "./scripts/post-plan.sh"
+    pre_apply  = "./scripts/pre-apply.sh"
+    post_apply = "./scripts/post-apply.sh"
+  }
 }`)
 }
 
@@ -374,6 +423,12 @@ resource "scalr_workspace" "test" {
   name           = "renamed-out-of-band"
   environment_id = scalr_environment.test.id
   auto_apply     = true
+  hooks {
+    pre_plan   = "./scripts/pre-plan.sh"
+    post_plan  = "./scripts/post-plan.sh"
+    pre_apply  = "./scripts/pre-apply.sh"
+    post_apply = "./scripts/post-apply.sh"
+  }
 }`)
 }
 
@@ -386,5 +441,29 @@ resource "scalr_workspace" "test" {
   operations            = false
   terraform_version     = "0.12.19"
   working_directory     = "terraform/test"
+  hooks {
+    pre_plan   = "./scripts/pre-plan_updated.sh"
+    post_plan  = "./scripts/post-plan_updated.sh"
+    pre_apply  = "./scripts/pre-apply_updated.sh"
+    post_apply = "./scripts/post-apply_updated.sh"
+  }
+}`)
+}
+
+func testAccScalrWorkspaceUpdateWorkingDir(rInt int) string {
+	return fmt.Sprintf(testAccScalrWorkspaceCommonConfig, rInt, defaultAccount, `
+resource "scalr_workspace" "test" {
+  name                  = "workspace-updated"
+  environment_id 		= scalr_environment.test.id
+  auto_apply            = false
+  operations            = false
+  terraform_version     = "0.12.19"
+  working_directory     = ""
+  hooks {
+    pre_plan   = "./scripts/pre-plan_updated.sh"
+    post_plan  = "./scripts/post-plan_updated.sh"
+    pre_apply  = "./scripts/pre-apply_updated.sh"
+    post_apply = "./scripts/post-apply_updated.sh"
+  }
 }`)
 }